Parsing

+++ sql/schema/002_programs.sql
@@ -2,7 +2,7 @@
 CREATE TABLE programs (
     id UUID PRIMARY KEY,
     name TEXT UNIQUE NOT NULL,
-    max_courses NUMERIC(1) NOT NULL
+    max_courses smallint NOT NULL
 );
 -- +goose Down 
 DROP TABLE programs;
